[PLAT-12658] Fixing edit universe to use correct config to fetch the universe namespace
Summary:
Fixing edit universe to use correct config to fetch the universe namespace
Prior to this change in EditUniverse workflow in volume resize code we were determining the universe namespace incorrectly.
This led to the failed check for needsExpandPVC, which skipped the PVC resize workflow.
Due to which the task failed.

This change updates the check to correctly use the azConfig instead that has the correct value of the namespace.

Test Plan:
{F149915}{F149907}
Tested by using the operator to create universe and edit its disk size, which completed.
Tested by using the UI to create a universe and edit its disk size which completed to validate existing behavior works.
Tested by using the UI to create a namespace per az and completed the disk size resize operation to ensure no regression.

Reviewers: sneelakantan, vkumar, sanketh

Reviewed By: sneelakantan, vkumar

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32385