[PLAT-10835]UI crashes on master if gflags api returns a 500 error
Summary:
Crash:
Previously, the UI required the "error" attribute to either be a string or an array of strings. However, if the backend returned an Object as an error, the UI crashed when attempting to access the "join" function of the Object. This is because "join" is only available for arrays, not Objects.

Fix:
To resolve this issue, we've updated the error message schema. Now, we first check if we have a "message attribute" (Object). If we do, we return it. Otherwise, we combine all the values of the array and return it. Additionally, the entire function is now enclosed in a try/catch block. If any errors occur, we simply return "Something went wrong" to prevent a crash.

Test Plan: Tested manaully.

Reviewers: lsangappa, jmak, rmadhavan

Reviewed By: lsangappa

Subscribers: ui, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32387