[PLAT-12676]: Fix the configure db apis for re-trigger
Summary:
If  the configure db APIs task fails after resetting the password while disabling the authentication, the subsequent re-trigger of the task will fail as the old password is changed to default.
So, we modified the validation in such a way that if it fails with the old password, we will try with the default password to check if it has already changed.

Test Plan: Tested manually by failing the disable ysql/ycql task after password change and verified that re-trigger succeeds.

Reviewers: sneelakantan, anijhawan

Reviewed By: anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32392