[PLAT-12416] Handle preview and deprecated v2 apis marked with x-yba-api-visibility
Summary:
This change adds the ability to mark API visibility on v2 API in the openapi spec. This is done by adding, for ex. `x-yba-api-visibility: preview` to the openapi spec. Valid values for `x-yba-api-visibility` are `preview`, `deprecated` and `internal` (same as the existing annotaiton `@YbaApi(visibility='xxx')`). The x-yba-api-visibility property can be set at various levels of the openapi spec - entire path (covers routes), entire schema (covers models for request and response), or properties within schema (covers individual properties of models).

#### Internal APIs
Whenever the API visibility is marked with `x-yba-api-visibility: internal`, these parts of the API are filtered out and only the remaining is written out to `src/main/resources/openapi_public.yaml`. This is used as input to generate the public YBA API docs in Stoplight.

#### Preview APIs
If API is marked with `x-yba-api-visibility: preview`, it is mandatory to also mark `x-yba-api-since: <yba_version>`, for example `x-yba-api-since: 2.20.0.0`, alongside the visibility. This is mandated by this script that is run during the sbt build. The script also automatically generates the corresponding preview warning message to the API description that will show up in the API docs.

#### Deprecated APIs
If API is marked with `x-yba-api-visibility: deprecated`, it is mandatory to also mark `x-yba-api-since: <yba_version>`, for example `x-yba-api-since: 2.20.0.0`, alongside the visibility. This is mandated by this script that is run during the sbt build. The script also automatically generates the corresponding deprecation warning message to the API description that will show up in the API docs.

Test Plan:
1. With visibility, but no since:
Manually added the following:
```
'/customers/{cUUID}/universes/{uniUUID}/upgrade/gflags':
  parameters:
    - name: cUUID
      <snip>
  post:
    operationId: upgradeGFlags
    summary: Upgrade GFlags
    x-yba-api-visibility: preview <---- added
```
Ran `sbt openApiBundle` and verified it fails:
```
Processing paths that are marked 'x-yba-api-vibility: deprecated' or 'preview'
Traceback (most recent call last):
  File "./openapi_generate_public.py", line 223, in <module>
    process_visibility_in_paths()
  File "./openapi_generate_public.py", line 131, in process_visibility_in_paths
    raise Exception(", ".join(errMsgs))
Exception: x-yba-api-since: is missing in post method of /customers/{cUUID}/universes/{uniUUID}/upgrade/gflags
[error] stack trace is suppressed; run last openApiBundle for the full output
[error] (openApiBundle) openapi bundle failed!!!
[error] Total time: 3 s, completed Feb 12, 2024, 8:45:51 AM
[yugaware] $
```

2. Preview visibility:
Manually tested by adding the following for preview:
```
'/customers/{cUUID}/universes/{uniUUID}/upgrade/gflags':
  parameters:
    - name: cUUID
    <snip>
  post:
    operationId: upgradeGFlags
    summary: Upgrade GFlags
    x-yba-api-visibility: preview <---- added
    x-yba-api-since: 2.21.0.0 <---- added
    description: 'Queues a task to perform GFlags upgrade in a universe. The input set of GFlags will replace any existing GFlags in the universe. Refer [YB-Master configuration flags](https://docs.yugabyte.com/preview/reference/configuration/yb-master/#configuration-flags) and [YB-TServer configuration flags](https://docs.yugabyte.com/preview/reference/configuration/yb-tserver/#configuration-flags).'
```

And ran `sbt openApiBundle` and verified that src/main/resources/openapi_public.yaml has this:
```
paths:
  /customers/{cUUID}/universes/{uniUUID}/upgrade/gflags:
    parameters:
    - name: cUUID
      <snip>
    post:
      operationId: upgradeGFlags
      summary: Upgrade GFlags
      x-yba-api-visibility: preview
      x-yba-api-since: 2.21.0.0
      description: '<b style="color:#FFA500">WARNING: This is a preview API in YBA
        version 2.21.0.0 that could change.</b></p>Queues a task to perform GFlags
        upgrade in a universe. The input set of GFlags will replace any existing GFlags
        in the universe. Refer [YB-Master configuration flags](https://docs.yugabyte.com/preview/reference/configuration/yb-master/#configuration-flags)
        and [YB-TServer configuration flags](https://docs.yugabyte.com/preview/reference/configuration/yb-tserver/#configuration-flags).'
```

3. Similarly, tested that `x-yba-api-visibility: deprecated` add the deprecation description.
4. And tested that `x-yba-api-visibility: internal" removes the object from openapi_public.yaml spec.

Reviewers: anijhawan, vpatibandla

Reviewed By: anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32354