[PLAT-12439] Make a single metrics call for table level replication lag metrics
Summary:
**Context**
The xCluster tab in the universe details page shows a warning banner with a count of all the tables exceeding the user determined max acceptable lag.
Prior to this diff, the UI would send a query to the `/metrics` endpoint for each table in the xCluster config.
To support xCluster configs with thousands of tables, we should instead send a single query to the `/metrics` endpoint asking for all the table metrics.

**Changes**
This diff sends a single query to the `/metrics` endpoint for `async_replication_sent_lag` and doesn't filter for a specific table.
The returned metrics data is then used for both the warning banner and the replication lag text shown per table.

This diff also contains minor text updates for 'full copy'/'bootstrapping'.

Test Plan:
Create a universe with 1000 tables and set up xCluster config for all those tables.
Verify that the client is only sending a single `/metric` request for `async_replication_sent_lag` every 15 seconds.
Verify that the response to the `/metric` request for `async_replication_sent_lag` contains replication lag for all tables in the
xCluster config.

Verify that the YBA UI displays replication lag per table and a warning banner when any tables exceed the user defined max acceptable lag.
{F148705}
{F148706}

Reviewers: rmadhavan, cwang, sanketh

Reviewed By: rmadhavan

Subscribers: kkannan, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32269