[PLAT-12417] Add x-yba-api-authz rbac extension to v2 apis
Summary:
This change adds ability to mark v2 APIs with a `x-yba-api-authz` object to describe the Authz requirement for the API path. This property is recognized only on a path element of the openapi spec. The structure of this object closely follows the `@AuthzPath` annotation that is applied on v1 controller methods.

With the change in this diff, the controller class now gets generated with the @AuthzPath annotation:

For a complete example:
```
    x-yba-api-authz:
      - requiredPermission:
          resourceType: universe
          action: UPDATE
        resourceLocation:
          path: universes
          sourceType: endpoint
          dbClass: com.yugabyte.yw.models.Backup.class
          identifier: backups
          columnName: backup_uuid
        checkOnlyPermission: true
```
generates this annotation for the method in the controller class:
```
  @AuthzPath({
    @RequiredPermissionOnResource(
        requiredPermission =
            @PermissionAttribute(resourceType = ResourceType.UNIVERSE, action = Action.UPDATE),
        resourceLocation = @Resource(path = Util.UNIVERSES,
          sourceType = SourceType.ENDPOINT,
          dbClass = com.yugabyte.yw.models.Backup.class,
          identifier = "backups",
          columnName = "backup_uuid"),
        checkOnlyPermission = true)
  })
```

Test Plan:
Added a user without the Universe.UPDATE permission.

Invoked the API:
```
curl --location 'http://localhost:9000/api/v2/customers/f33e3c9b-75ab-4c30-80ad-cba85646ea39/universes/1c8568cf-88cd-4e84-9df4-f7fedad4eaaf/upgrade/gflags' \
--header 'Content-Type: application/json' \
--header 'Accept: application/json' \
--header 'X-AUTH-YW-API-TOKEN: 6ccdb012-806b-4fd5-b52f-df253e624498' \
--data '{
  "primary_gflags": {
    "per_process_gflags": {
      "tserver_gflags": {
            "allow_insecure_connections": "true",
            "enable_stream_compression": "true"
      },
      "master_gflags": {
        "allow_insecure_connections": "true"
      }
    }
  }
}
'
Response: Unable to authorize user
```

Verified that the API failed with this authz log:
```
2024-02-12T09:58:10.691Z  [debug] 9ff62564-9349-4a9b-8cb5-69fd0f96ac6a AuthorizationHandler.java:121 [application-pekko.actor.default-dispatcher-1560] com.yugabyte.yw.rbac.handlers.AuthorizationHandler User f3dd807e-3950-4cd1-860f-548bb581dc05 does not have the required permission UPDATE on the resource UNIVERSE
```

Reviewers: vpatibandla, #yba-api-review!

Reviewed By: vpatibandla

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32356