[pg15-upgrade] fix: avoid assertion in debug build
Summary: Testing of the PG11 to PG15 upgrade has been on the release build, because there is no public download of a debug build, and the build flavors are incompatible at the storage layer. When doing an upgrade from a debug build of PG11 to a debug build of PG15 (which is not currently supported), an assertion fires. The assertion doesn’t indicate a bug, so fix the code to make an equivalent check that doesn’t assert.

Test Plan:
Jenkins: rebase: pg15-upgrade, compile only

Make the following temporary change to src/yb/common/column_id.h to allow debug build to read release build storage:
    @@ -95,7 +95,7 @@ static const ColumnId kInvalidColumnId = ColumnId(std::numeric_limits<ColumnIdRe
     #ifdef NDEBUG
     constexpr ColumnIdRep kFirstColumnIdRep = 0;
     #else
    -constexpr ColumnIdRep kFirstColumnIdRep = 10;
    +constexpr ColumnIdRep kFirstColumnIdRep = 0;
     #endif
     const ColumnId kFirstColumnId(kFirstColumnIdRep);

(Alternatively, build a PG11 fastdebug at the same commit as the YB master base of the pg-upgrade branch, and modify pg15_tests/test_upgrade.sh to upgrade from that.)

On MacOS 13 arm64:
    ./yb_build.sh fastdebug --sj
    pg15_tests/test_upgrade.sh

The upgrade will still fail, but it's on a later assertion that has a more involved fix, yet to be sent out. You will see these errors in one of the /tmp/pg15_cluster_data/node-?/disk-1/master.err files.

Original assertion:
    entity_ids.cc:159] Check failed: (pos == 2) && (variant & 0xC0) == 0x80 Invalid Postgres id 00000000000000000000000000000003

Assertion with this fix:
    column_id.cc:22] Check failed: t >= 0 (-8 vs. 0)

Reviewers: jason, hsunder

Reviewed By: jason, hsunder

Subscribers: ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32433