Merged PR 1274312: [Index] Improve ComparePartial performance
### Does this PR have any customer impact?
Yes

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)
Feature

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)
No

### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description
When doing compare partial today, the index layer does a tuple store put/get/sort. This means it's sensitive to the tuple size which today is a full scan item. In the case where useAlternativeOrder and useCurKey is false (no sorts) we don't need as much information. Make the tuple store smaller in these cases.

Based on testing on simple scenarios of $exists of a few million documents, we see it go from ~2s to 1.4s for a compare partial query.