[PLAT-12681] Handling Azure caching limitations
Summary: Issue with ultra disks where we were trying to set caching to ReadOnly when it was not supported. Also have run into sizing issues in the past, so this diff removes the setting of caching entirely. With updates to our Azure API (or a fix on their end) sending disk creation without caching information automatically picks up ReadOnly for data disks whenever possible, and None for ultra disks or disks too large.

Test Plan: Create VMs with standard/premium/ultra and different sizes (100 or 5000). Ensure that caching property is set to ReadOnly whenever possible and None otherwise.

Reviewers: yshchetinin, svarshney, nbhatia, rohita.payideti

Reviewed By: rohita.payideti

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32447