[#20912] DocDB: Use rpc::Strand instead of thread pool token in PeerMessageQueue
Summary:
rpc::Strand is the equivalent of ThreadPoolToken but for use with rpc::ThreadPool. To avoid slowness with condition variable signaling that is causing a significant TPCC NewOrder latency increase, use a Strand within a dedicated rpc::ThreadPool instead of ThreadPoolToken in PeerMessageQueue::NotifyObserversOfMajorityReplOpChange and PeerMessageQueue::NotifyObservers. The respective rpc::ThreadPool has unbounded threads (creating new threads as needed without a limit) and lives inside TSTabletManager (or sys catalog in case of master).

As is necessary for the above, adding support for an unbounded number of threads in rpc::ThreadPool, mainly by not reserving space in the vector of workers.

Adding thread safety analysis annotations to rpc::ThreadPool and making sure the vector of workers is always protected by the mutex.

Also changing the logic of creating new workers in rpc::ThreadPool. In case of a failure to create a worker, we will not increment the worker counter, so that we would try again next time.

Also fix handling of ~/.yb_buildrc and .git/yb_buildrc files in yb_build.sh.
Jira: DB-9894

Test Plan:
Jenkins

TPCC:
- https://github.com/yugabyte/yugabyte-db/issues/20912#issuecomment-1945493060
- https://github.com/yugabyte/yugabyte-db/issues/20912#issuecomment-1945495030
- https://github.com/yugabyte/yugabyte-db/issues/20912#issuecomment-1945525578

Reviewers: sergei

Reviewed By: sergei

Subscribers: ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32353