[#20438] YSQL: Scalar array operation pushdown
Summary:
Add support for pushdown of ScalarArrayOpExpr expression nodes.
ScalarArrayOpExpr has two arguments, one is a scalar, other is an array.
The DocDB side of support iterates over elements in the array and
applies the operation, a two argument function, which takes the scalar
and the next element of the array on each iterations. The function
results are either OR'ed or AND'ed, as defined by the useOr flag.

The iteration over array elements is based on Postgres'
deconstruct_array function which requires some information about element
data type to find specific array elements, in particular, type data length,
either fixed or variable, is it by-value or by-reference, alignment.
This information is available in system catalog (pg_type system table), but
since catalog is not accessible from within DocDB we introduce a
function to return those type attributes by type id from static data.
Naturally, only SAOPs on arrays where element types are supported by
that function are allowed to be pushed down. An unit test is added to
make sure the statis data matches the data in the catalog.
Jira: DB-9438

Test Plan: ybd --cxx-test=pg_type-test

Reviewers: dsrinivasan, gkukreja, skumar, tnayak

Reviewed By: tnayak

Subscribers: ybase, smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D31588