[#21058] DocDB: Revert "[#14165] DocDB: Rollback only the newest transaction in a deadlock"
Summary:
This reverts commit cc1a9c061e51d2030daeca9c34638119f2d18320.

The above commit caused a regression for contentious serializable workloads.

Earlier, when deadlock cycles were detected, we used to abort the transaction that originated the probe. But the above commit changes that logic to abort the youngest transaction instead. The problem with this approach is that when the origin transaction is part of two or more deadlock cycles, our deadlock algorithm ignores the callback for all cycles except the first detected cycle. So with the new logic, we end up in a deadlocked state, but don't detect it, leading to requests being timed out in the wait-queue (and new requests coming in, which trigger the probing algorithm again, which is when another deadlock might be detected). Despite the algorithm ignoring callbacks for subsequent cycle detected for a particular probe, we don't have this problem with the earlier algorithm since it aborts the originating transaction, breaking all possible deadlock cycles.

The test `testSerializableReadDelayWrite` rightly catches this issue.

**Upgrade/Downgrade safety** - The above commit hasn't been part of any official release yet, so reverting the new proto changes for now. This shouldn't have any implications on upgrades/downgrades among stable releases.
Jira: DB-10025, DB-3646

Test Plan: ./yb_build.sh --java-test 'org.yb.pgsql.TestPgTransactions#testSerializableReadDelayWrite'

Reviewers: rthallam, rsami, hsunder

Reviewed By: rthallam, rsami

Subscribers: ybase, yql, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32461