[#21014] DocDB: Add clone support in import_snapshot
Summary:
This diff adds support for cloning in `DoImportSnapshotMeta`. It is only called from a test at this point. If `is_clone` is specified, the function will always create the tablets instead of only if repartitioning is required. The tablets are created in the CREATING state, and will be created on tservers by the clone source tablets applying a clone op.

**Upgrade/Rollback safety:**
The proto features used in this diff are only called from tests. They will eventually be guarded by an auto flag.

Jira: DB-9988

Test Plan: `ybd --cxx-test backup-txn-test --gtest_filter BackupTxnTest.ImportSnapshotMetaCloneYcql`

Reviewers: mhaddad, xCluster, hsunder

Reviewed By: mhaddad

Subscribers: loginov, ybase, bogdan

Differential Revision: https://phorge.dev.yugabyte.com/D32476