[PLAT-12656] Add autoflag validation prechecks for DR and xcluster
Summary:
Add extra autoflags validation on DR operations: create, switchover, change replica, restart, and add/drop table. We also add validation for xcluster creation and restart.

For DR create, we validate that all promoted/modified autoflags on primary must also be promoted/modified on the standby and vice versa. This is because on switchover, the primary and standby roles are switched and we would like to reduce customer confusion (i.e. on setup, it is fine for the standby universe to have more promoted autoflags than the primary universe).

In addition, this fixes the issue where one of the universes, say universe A is initially created with db version 2.20.2 while the other universe, say universe B was on an older version, say 2.16 and then upgraded to 2.20.2. For the universe that was created with db version 2.20.2 will have `ysql_enable_packed_row` autoflag enabled while the other universe that was upgraded to 2.20.2 will not have `ysql_enable_packed_row` autoflag enabled. Then, if we set up DR with A -> B, it will fail.

Add runtime config, `yb.xcluster.enable_auto_flag_validation` to enable/disable these checks as needed

Test Plan:
Test scenarios

1. Test original pre-check on DR setup. Create a universe A with db version >=2.20.2. Create a universe B with db version <= 2.18.0.0 and then upgrade to 2.20.2. Create DR with A -> B and also B -> A. Check that both will fail saying that `ysql_enable_packed_row` is not enabled on one of the universes.
2. Create universe A and B 2.20.2 db version. Set up DR A -> B. Create universe C with db version 2.17.1.0-b300 and then upgrade to 2.20.2. Run replace replica task, i.e. try to re-setup DR A -> C. make sure that the validation fails and we cannot setup A -> C

Reviewers: sanketh, jmak

Reviewed By: sanketh

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32379