[PLAT-12450] Replace system level systemd checks with user level
Summary:
In our code, we use to check for system or user level systemd by reading files at path prefix
`/etc/systemd/system`. Now, it could be the case that for manually provisioned nodes we might not
have the read permissions on that path.
This diff changes that check to read from user level systemd path prefix.

Test Plan:
Created a univrse with alma8.
Verified we are using user level systemd.

Created a centos universe using 2.19 build.
Verified that we are using system level systemd.

Reviewers: sanketh, muthu

Reviewed By: muthu

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32357