[#21087] CDCSDK: Fix TestCleanupAfterLateAlterTable by using correct sleep time
Summary:
TestCleanupAfterLateAlterTable has been failing because I set the DDL timeout to a small value in the test. Creation of database is a slow operation, so the test times
out in most iterations.

This revision fixes it by removing the small value and just using the default value.
Jira: DB-10048

Test Plan:
Jenkins: test regex: .*TestCleanupAfterLateAlterTable.*

```
./yb_build.sh --cxx-test cdcsdk_consistent_snapshot-test --gtest_filter CDCSDKConsistentSnapshotTest.TestCleanupAfterLateAlterTable -n 10
```

Reviewers: asrinivasan, devansh.saxena

Reviewed By: asrinivasan, devansh.saxena

Subscribers: ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D32484