[#17132] DocDB: Exit early when create table fails and return failure error
Summary:
Return early if table creation failed and include the failure error in the response.
Currently any table create error after a certain point is treated as a error in the retrieval of the IsCreateTableDone RPC. So we need to wait for the DDL to timeout and it returns a generic "Timed out" error.
With this change we will set done on the response and the client will stop the retries immediately and return the actual error to the client.

Fixed a bug in AddTableToXClusterTargetTask where we were binding a function returning a Status to a void causing the error to get missed.

Fixes #17132
Jira: DB-6413

Test Plan: xcluster_ysql_index-test

Reviewers: asrivastava, jhe, xCluster

Reviewed By: asrivastava, jhe

Subscribers: bogdan, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D32474