[PLAT-12641] Block adding instance type with existing instance type code
Summary:
**Context**
The YBA UI has an 'Add instance type' modal which actually performs an edit
operation (due to the backend using `upsert`) when the user provides an instance
type name which is already in use.

**Change**
This diff adds a validation step which blocks the user from submitting instance
type names which are already in use to avoid an unintended edit operation.

Test Plan:
Verify that the user is no longer able to submit an already use instance type name
when in the 'add instance type' modal.

Reviewers: rmadhavan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32507