[PLAT-12682][PLAT-12601][PLAT-12698] Update DR text and error messages
Summary:
**Context**
We want to start referring to DR and xCluster both as xCluster in our user facing messaging.
DR will be referred to as xCluster DR and xCluster will be referred to as xCluster replication.

**Changes**
- Update xCluster and DR tab title and header.
- Add/modify error messages
- Move DR metrics link to separate tab

Test Plan:
Verify text changes.
{F150250}
{F150251}
Verify DR metric link is in a separate tab.
{F150252}
Verify error messages.
{F150249}
{F150457}
{F150463}
(Changed term to 'database' from 'table' for DR and ysql xCluster)

Reviewers: rmadhavan, kkannan, lsangappa

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32419