[#15718][CDCSDK] Fix for tablet split data loss
Summary:
In the current implementation, when explicit checkpointing is enabled, due to buggy implementation the `cdc_sdk_safe_time` is moved forward to the `safe_hybrid_time` with every get changes request. This means that we would update the `cdc_sdk_safe_time` without receiving an explicit checkpoint confirming the consumption of records corresponding to previous safe time. This behaviour could lead to data loss when a tablet is split. The `getTabletListToPollForCDC()` returns an Op Id corresponding to the last checkpointed record on the parent tablet. However the checkpointed safe time is incorrect and could lead to filtering of records.

This diff introduces a fix for this problem by increasing the `cdc_sdk_safe_time` only when an acknowledgement is received from the GetChanges request explicit checkpoint.

Test Plan: Existing cdc tests and QA tests

Reviewers: skumar, vkushwaha, asrinivasan

Reviewed By: asrinivasan

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32455