[BACKPORT 2.20][#15718][CDCSDK] Fix for tablet split data loss
Summary:
Original commit: 9246120dc5178cd38c31f24ad3ac7966910931d9 / D32455
In the current implementation, when explicit checkpointing is enabled, due to buggy implementation the `cdc_sdk_safe_time` is moved forward to the `safe_hybrid_time` with every get changes request. This means that we would update the `cdc_sdk_safe_time` without receiving an explicit checkpoint confirming the consumption of records corresponding to previous safe time. This behaviour could lead to data loss when a tablet is split. The `getTabletListToPollForCDC()` returns an Op Id corresponding to the last checkpointed record on the parent tablet. However the checkpointed safe time is incorrect and could lead to filtering of records.

This diff introduces a fix for this problem by increasing the `cdc_sdk_safe_time` only when an acknowledgement is received from the GetChanges request explicit checkpoint.

**Backport Description**
The test CDCSDKBeforeImageTest.TestCompactionWithConsistentSnapshotAndNoBeforeImage failed in 2.20 and 2.20.2 but not on master. This was because the changes introduced by https://github.com/yugabyte/yugabyte-db/commit/9d47f24cb8857b708a8de993db5c7b158e48157 are not present in these branches. As a result when the test calls GetChanges by providing an explicit checkpoint with snapshot time set to zero, the cdc_sdk_safe_time in the cdc_state table was being reset to Invalid. To fix this, if an empty optional value is passed for safe_time to the method UpdateCheckpointAndActiveTime(), then we get the checkpoint from the state table and rewrite the safe time with the safe time from the checkpoint. Since the connector is not expected to send an explicit checkpoint without snapshot_time this code is not expected to be hit in normal scenarios.

Test Plan: Existing cdc tests and QA tests

Reviewers: skumar, vkushwaha, asrinivasan

Reviewed By: asrinivasan

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32525