[PLAT-12779] Handle undefined source universe in table details metric page
Summary:
The sourceUniverse prop is possibly undefined in `CustomerMetricsPanel` which means we should be optionally indexing `drConfigUuidsAsSource` and `drConfigUuidsAsTarget` to avoid an error from indexing undefined.
This diff resolves this issue by adding the optional chaining operator to resolve to `undefined` instead.

Test Plan:
Verify that the table details pages load correctly.
{F151534}
{F151535}
Verify that the universe metrics page loads fine and all the tabs work as expected.
{F151536}

Reviewers: rmadhavan

Reviewed By: rmadhavan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32533