[BACKPORT 2.20][21004] YSQL: Skip computation of range bounds while pre-checking Index-Only Scan recheck condition
Summary:
Original commit: 59bc43f786c63b0faef3ca382b5f06165a155d4f / D32348
In a query of the form:
`SELECT <expression> FROM table WHERE table.col <= <cond1> AND table.col <= <cond2>` , we seek to compute the minimum of (<cond1>, <cond2>) in order to simplify the query to `SELECT <expression> FROM table WHERE table.col <= min(<cond1>, <cond2>)` so that the result can be computed using a single filter. This is applicable when there are two or more (`<, <=` or `>, >=`) operators.

This computation is performed in the following scenarios:

1. The beginning of a scan (or rescan) (`ybcBeginScan`)
2. As a precheck of the recheck condition for aggregate pushdown

This computation can be skipped during precheck because:
1. It causes a crash for certain queries: In a correlated subquery, when one of the conditions in the above query is correlated to the outer query, the values of the bounds is not available for calculation of the minima/maxima in the precheck stage (as the outer table would not yet have been scanned).

2. It is unnecessary: The calculated range bounds has no influence during precheck and is discarded soon after.

Test Plan:
```
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressIndex#testPgRegressIndex'
```

Reviewers: tnayak, jason, mtakahara

Reviewed By: jason

Subscribers: smishra, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D32545