[pg15] test: add yb_dep_create_index
Summary:
Many tests depend on certain objects created by test_setup.  They may
also depend on indexes for those objects, typically created by the
create_index test.  Although this amount of dependencies is an
improvement over PG 11, it is still time-consuming to run both
test_setup and create_index tests.  That is why yb_dep_* tests exist to
provide individual objects.  But coverage of those individual objects is
currently limited to only the most frequent ones, and it can be a hassle
to find all the individual objects in case there are many.

That is the case with the aggregates test.  It has multiple
dependencies, some which do not yet have a yb_dep_* test, such as
varchar_tbl.

Come up with a new mega-dependency yb_dep_create_index which is the same
as create_index but should be much faster by

- skipping unnecessary queries
- making all CREATE INDEX nonconcurrent

The yb_pg_aggregates test still fails, but there are far fewer
differences compared to before thanks to providing the indexes.  A
followup is planned to get that test fully passing.

Test Plan:
On Almalinux 8:

    pg15_tests/test_aggregates.sh fastdebug --gcc11

Jenkins: skip

Reviewers: aagrawal, tfoucher, tvesely

Reviewed By: tfoucher

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D32548