[PLAT-12791][Platform]Fix Linux version for Read replica
Summary:
Before, when displaying the Linux version used for a read replica, we got the information from the primary cluster. This was because through the UI, we could only generate the read replica with the same provider used for the primary.

However, through the API, we have the capability to use a different provider when creating the read replica. so, we have now adjusted the code to utilise the appropriate cluster for retrieving the Linux version.

Test Plan: Tested manually

Reviewers: lsangappa, svarshney

Reviewed By: lsangappa

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32560