[PLAT-12773] Fixes to provider validation failure response
Summary:
Move the logic to add data prefix to
CustomerConfigStorageValidator. It is not needed in case of other
validations as only the CustomerConfig has all the required things
under the data key.
- AWS provider validation no longer has this data. prefix now, so
  update the UI logic accordingly.

Errors related to a field are now passed on as multiple errors in the
array instead of returning an array with single comma separated
errors.

Add requestJson field which holds the client supplied request
payload. This is added to YBPStructuredError and YBPError.

Test Plan:
1. Tried creating Kubernetes provider with mock validation enabled,
   the data prefix is no longer present and the request JSON is present
   in the API response.

2. Tried creating AWS provider with invalid access credentials, the UI
   works as expected with fields not having data prefix.

3. Tried creating a storage config with wrong GCP credentials, the
   returned error has data prefix as expected.

Reviewers: anijhawan, dshubin, lsangappa, #yba-api-review

Reviewed By: anijhawan, dshubin, #yba-api-review

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D32518