[#17482] YSQL: Tablespace with colocated tables
Summary:
Currently, we do not have the ability to specify the placement policy of colocated tables in a colocation database. Colocated tables in a database are located on a single tablet whose replicas are distributed throghout the cluster. To specify placement policy of colocated tables, we will allow use of tablespaces with colocated tables, as we do with non-colocated tables. This would require placing colocated tables in a colocated tablet, based on their tablespace. Instead of having a single implicit tablegroup (and hence a single colocated tablet) per db, we will now have a separate implicit tablegroup for each tablespace used by a colocated table.

For example, if we have 4 colocated tables (t1, t2, t3, t4) with tablespaces (ts1, ts1, ts2, ts3), we should have 3 colocated tablets (and therefore 3 implicit tablegroups) in the colocation database, corresponding to (ts1, ts2, ts3).

    t1, t2 -> implicit tablegroup corresponding to ts1 -> colocated tablet 1
    t3 -> implicit tablegroup corresponding to ts2 -> colocated tablet 2
    t4 -> implicit tablegroup corresponding to ts3 -> colocated tablet 3

When we receive a CREATE TABLE request for a colocated table with tablespace specified, we shall check in pg_shdepend for a dependency from the given tablespace to a tablegroup in the same DB. If such an entry does not exist we shall create a new implicit tablegroup having the specified tablespace and insert the dependency entry to pg_shdepend. Further CREATE TABLEs (for colocated tables with the same tablespace) will use this implicit tablegroup.

We shall drop the implict tablegroup when it has no tables. When a DROP TABLE is executed, we shall check for dependencies to the implicit tablegroup (if any). If no other dependencies exist, we will drop the implicit tablegroup.

We will allow the creation of colocated indexes with tablespace independently of the base table, i.e., we can explicitly set tablespace for indexes. If an index does not specify a tablespace in a colocated database, then it is assigned to the pg_default tablespace and placed in the default tablegroup. This is similar to the behaviour for indexes of non-colocated tables.

We allow backups for all cases, and using the '--use-tablespaces' flag will restore the collocated tablets to the same topology with the tablespaces being recreated. However, for the case when the '--use-tablespaces' flag is not employed, then on the restore, if the same topology exists, we’ll restore to these tablespaces as usual. If the topology does not exist, we’ll restore colocated tablets to the pg_default tablespace.

[[ https://docs.google.com/document/d/15t_6QxZyIpDvBqJkQ1ryXGLyy3VRmsHNBVRg1C96Wg4/edit# | Functional Spec ]]

Test Plan:
yb_build.sh --cxx-test colocation-test --gtest_filter ColocatedTablesWithTablespacesTest.ColocatedTablespaceTest
yb_build.sh --cxx-test colocation-test --gtest_filter ColocatedTablesWithTablespacesTest.ColocatedIndexWithTablespaceTest
yb_build.sh --cxx-test colocation-test --gtest_filter ColocatedTablesWithTablespacesTest.ColocatedPartitionedTableWithTablespaceTest
yb_build.sh --cxx-test yb-backup-cross-feature-test --gtest_filter YBBackupTestColocatedTablesWithTablespaces.TestBackupColocatedTablesWithTablespaces
yb_build.sh --cxx-test yb-backup-cross-feature-test --gtest_filter YBBackupTestColocatedTablesWithTablespaces.TestBackupColocatedTablesWithoutTablespaces
yb_build.sh --java-test org.yb.pgsql.TestPgRegressColocatedTablesWithTablespaces#testPgRegressColocatedTablesWithTablespaces

Reviewers: skumar, tverona, yguan, stiwary

Reviewed By: skumar, yguan

Subscribers: yql, ybase, bogdan, yguan, tverona

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D24681